requirements.txt
# 音声処理
faster-whisper>=0.10.0
ffmpeg-python>=0.2.0

# テキスト処理
pandas>=1.5.0
unidic-lite>=1.0.8
fugashi>=1.3.0

# AI・API
openai>=1.0.0

# 設定管理
python-dotenv>=1.0.0

# Web API
flask>=2.3.0
flask-cors>=4.0.0

# データベース
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pgvector>=0.2.0

# Google Cloud
google-cloud-storage>=2.10.0
google-cloud-sql-connector>=1.4.0

# メッセージング
google-cloud-pubsub>=2.18.0
google-cloud-tasks>=2.14.0

# ログ・監視
google-cloud-logging>=3.8.0
google-cloud-error-reporting>=3.4.0

# その他
pydantic>=2.0.0
